refactor: custom hook useInitials
