+++ app/routes/_index.tsx
@@ -27,7 +27,6 @@ export default function Index() {
 
   return (
     <>
-      <Avatar userName="Jose Octavio" />
       <h1>Bienvenido. {user?.displayName ? `${user.displayName}.` : ""} </h1>
       <section
         className="grid gap-2"
